buffered newtasks input, merging duplicates with max budget. prevents duplicate task emits which actually occurr quite frequently

+++ on/src/main/java/nars/config/Parameters.java
@@ -77,37 +77,37 @@ public class Parameters {
     
     /* ---------- budget thresholds ---------- */
     /** The budget threshold rate for task to be accepted. */
-    public static final float BUDGET_THRESHOLD = (float) 0.01;
+    public static final float BUDGET_THRESHOLD = 0.01f;
 
     /* ---------- default input values ---------- */
     /** Default expectation for confirmation. */
-    public static final float DEFAULT_CONFIRMATION_EXPECTATION = (float) 0.6;
+    public static final float DEFAULT_CONFIRMATION_EXPECTATION = 0.6f;
     /** Default expectation for creation of concept. */
-    public static final float DEFAULT_CREATION_EXPECTATION = (float) 0.66; //0.66
+    public static final float DEFAULT_CREATION_EXPECTATION = 0.66f; //0.66
      /** Default expectation for creation of concept for goals. */
-    public static final float DEFAULT_CREATION_EXPECTATION_GOAL = (float) 0.6; //0.66
+    public static final float DEFAULT_CREATION_EXPECTATION_GOAL = 0.6f; //0.66
     /** Default confidence of input judgment. */
-    public static final float DEFAULT_JUDGMENT_CONFIDENCE = (float) 0.9;
+    public static final float DEFAULT_JUDGMENT_CONFIDENCE = 0.9f;
     /** Default priority of input judgment */
-    public static float DEFAULT_JUDGMENT_PRIORITY = (float) 0.8;
+    public static float DEFAULT_JUDGMENT_PRIORITY = 0.8f;
     /** Default durability of input judgment */
-    public static float DEFAULT_JUDGMENT_DURABILITY = (float) 0.5; //was 0.8 in 1.5.5; 0.5 after
+    public static float DEFAULT_JUDGMENT_DURABILITY = 0.5f; //was 0.8 in 1.5.5; 0.5 after
     /** Default priority of input question */
-    public static final float DEFAULT_QUESTION_PRIORITY = (float) 0.9;
+    public static final float DEFAULT_QUESTION_PRIORITY = 0.9f;
     /** Default durability of input question */
-    public static final float DEFAULT_QUESTION_DURABILITY = (float) 0.9;
+    public static final float DEFAULT_QUESTION_DURABILITY = 0.9f;
 
     
      /** Default confidence of input goal. */
-     public static final float DEFAULT_GOAL_CONFIDENCE = (float) 0.9;
+     public static final float DEFAULT_GOAL_CONFIDENCE = 0.9f;
      /** Default priority of input judgment */
-     public static final float DEFAULT_GOAL_PRIORITY = (float) 0.9;
+     public static final float DEFAULT_GOAL_PRIORITY = 0.9f;
      /** Default durability of input judgment */
-     public static final float DEFAULT_GOAL_DURABILITY = (float) 0.9;
+     public static final float DEFAULT_GOAL_DURABILITY = 0.9f;
      /** Default priority of input question */
-     public static final float DEFAULT_QUEST_PRIORITY = (float) 0.9;
+     public static final float DEFAULT_QUEST_PRIORITY = 0.9f;
      /** Default durability of input question */
-     public static final float DEFAULT_QUEST_DURABILITY = (float) 0.9;
+     public static final float DEFAULT_QUEST_DURABILITY = 0.9f;
  
     
     /* ---------- space management ---------- */
@@ -274,9 +274,9 @@ public class Parameters {
     public static int TASKLINK_PER_CONTENT = 4; //eternal/event are also seen extra
     
     /** Default priority of exection feedback */
-    public static float DEFAULT_FEEDBACK_PRIORITY = (float) 0.9;
+    public static float DEFAULT_FEEDBACK_PRIORITY = 0.9f;
     /** Default durability of exection feedback */
-    public static float DEFAULT_FEEDBACK_DURABILITY = (float) 0.5; //was 0.8 in 1.5.5; 0.5 after
+    public static float DEFAULT_FEEDBACK_DURABILITY = 0.5f; //was 0.8 in 1.5.5; 0.5 after
     
     /* Priority penalty for sequences that start with an operation */
     public static float OPERATION_SEQUENCE_START_PENALTY = 0.05f;
